refactor: 랭킹 스케줄러 위치 변경, 감쇄 계수 낮추어 수정
