+++ apps/commerce-api/src/main/java/com/loopers/domain/ranking/RankingService.java
@@ -7,7 +7,6 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ZSetOperations;
 import org.springframework.stereotype.Service;
 
-import java.time.Duration;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Set;
@@ -32,27 +31,6 @@ public class RankingService {
                         t.getScore()
                 ))
                 .toList();
-    }
-
-    public void carryOverYesterdayRankings(LocalDate from, LocalDate to, double weight) {
-        String fromKey = RedisKeyGenerator.buildKey("ranking:all:", from);
-        String toKey = RedisKeyGenerator.buildKey("ranking:all:", to);
-
-        // 오늘 랭킹
-        Set<ZSetOperations.TypedTuple<String>> tuples =
-                redisTemplate.opsForZSet().rangeWithScores(fromKey, 0, -1);
-
-        if (tuples == null || tuples.isEmpty()) {
-            return;
-        }
-
-        // 가중치 적용하여 내일 랭킹에 추가
-        tuples.forEach(t -> {
-            double newScore = t.getScore() * weight;
-            redisTemplate.opsForZSet().add(toKey, t.getValue(), newScore);
-        });
-
-        redisTemplate.expire(toKey, Duration.ofDays(2));
     }
 
 }
